Refactor PostQuantumIntegrationTests and PostQuantumNet10Tests to use updated key generation and signing methods; remove obsolete RC4 tests; enhance Rabbit, XSalsa20, and ShamirSecretSharing tests for consistency and clarity; improve security hardening tests with better parameter validation.

+++ src/HeroCrypt/Cryptography/Primitives/Kdf/ScryptCore.cs
@@ -440,7 +440,7 @@ internal static class ScryptCore
     private static int Integerify(ReadOnlySpan<byte> block)
     {
         var offset = block.Length - 64;
-        return (int)(block[offset] | (block[offset + 1] << 8) | (block[offset + 2] << 16) | (block[offset + 3] << 24));
+        return block[offset] | (block[offset + 1] << 8) | (block[offset + 2] << 16) | (block[offset + 3] << 24);
     }
 
     /// <summary>
